New design of books for easier visual recognition
Curse books: Red colors and greish-tones to signify cursed.
Curse of Vanishing: Book is slightly seethru as it's phasing away.
Efficiency: Small book to indicate the efficiency of reading and faster read time etc.
Knockback: It's a brick. It literally is a brick in a book.
Fortune: Piles of coins! I think it might make sense.
Mending: Going with the original proposed design of having EXP bubbles.
Looting: Literal loot from minecraft to signify it's meaning plus a green-ish color to tie it with nature and mobs.
Loyalty: Colors of the Trident as it's a trident locked enchantment on Minecraft.
Mending: XP Orbs
Thorns: Spines/roots coming out of the book