Hermit palettes (#165)
* working on adding palettes, as well as the cub pharaoh card

pharaoh should work properly soon, but can be commented out like dream if it isn't to be used in the main game. However, having the card to use as an alternate palette is useful for testing which is why it's on this branch.

* palettes have been switched to functions, and the name length function has been fixed

finishing palettes

the palettes have been finalized, text should be working properly, and alter ego backgrounds are all one file. in addition, pharaoh has been removed from this branch.

* Rebase onto feature/everything-refactor

Co-Authored-By: Sense101 <67970865+Sense101@users.noreply.github.com>

---------

Co-authored-by: Sense101 <personal.sense101@gmail.com>
Co-authored-by: Sense101 <67970865+Sense101@users.noreply.github.com>