Added EX Boss and fixed minor visuals
Added boss variants of server logic to run game and a new button to matchmaking menu for creating a boss match.

EX is a new hermit card in a disabled "boss" expansion that cannot be put in regular play and has a unique status to handle a special ninth turn move.