Reorganize dependencies (#551)
This PR removes unused dependencies and moves the dev dependencies into the right place in the package.json. Additionally I removed every dependency that wasn't used in the packages CI.

If someone was relying these dependencies for a personal workflow, you should install the package globally as not to muck up the project.