Add CI To Release With Github Releases and Docker (#570)
This should make releases harder to mess up, theoretically. Tags will
allow us to only have one main branch, this means PR could never be
merged into the wrong place. I've used this system on a few projects in
the past and it makes releases very easy.
Additionally also improves the version display on the main menu to show
the commit. This would help with people running development builds and
understanding if they are on the right version.

Requires EVNs to be set up and there to be a dockerhub.