Change how armor stand uses discard (#576)
This is a huge issue in the code right now. Armor stand discarding
itself allows for a `RowStateWithHermit` to NOT have a Hermit. With this
change, it always will.

Now, only Hermit cards give a point when knocked out to compensate. I
think this is fine. If we need a different system, we can discuss that
in the future

I also simplified the attack loop, so attacks are NOT ran if they're
added after the main attack loop is over. Thorns ect. uses
`executeExtraAttacks` instead now.