Create status effect types and log fixes (#602)
- Status effects now have 4 types `normal`, `damage`, `system`, and
`hidden`. This affect some of their properties.
- Logs also work properly now. Before, apply logs needed to be specified
manually, and remove logs showed for all logs. Now the system is the
same as the system cards and attacks use, so it's a lot easier to follow
and works properly. Status effect logs can also be null for those that
shouldn't make logs.