Switch to Entity-Component Model (#604)
This PR switches the game over to the entity component mode. The
advantage of this model is that we can write queries to find exactly
what item we need. This maps very well to card descriptions.
- Player Component: Players are moved to the ECS to make searching and
modifying them more consistent with the rest of the game.
- Observer Component: To observe a hook, you must use the observer
object passed into the function. This gives us more control over when
hooks are removed, removing bugs such as the recent golden axe bug where
a hook wasn't removed. If you need a hook that lasts between turns the
proper way to do it is with status effects. Status effects also have
better telegraphing for the player.

# Advent of TCG
This PR comments out Advent of TCG because I want this system to be done
and testing before all the new alter egos are released so we can
implement them in the best possible way.

# Bug Fixes
- Fixed a bug where knockback would have broken if you killed an AFK Hermit while using knockback
- Converted multi-turn effects into status effects to make them easier to understand

---------

Co-authored-by: Alice <63879236+Benji42@users.noreply.github.com>