Fix Ren and Cleo sometimes not creating attacks (#613)
Fixes a bug where Ren and Cleo would not create a attacks when copying a
hermit. This was because they did not call the on attack request hook. I
now use `callSome` to call the specific `getAttackRequests` hook for the
created mock card before the attack loop.
Additionally I also prevented blocked actions from being selected
because this led to buggy behavior.