Fix Joehills/Cleo interaction and other unintuitive interactions (#668)
When Timeskip is used, it now blocks the secondary for all Joehills the
player has on the board. This fixing some unintuitive interactiosn with
Timeskip being able to be used but it not being able to fix a turn. This
also fixes the interaction with Cleo, where her secondary would be
blocked instead of Joe.