Add hotkey system, improve UI on mobile, and change muting system (#713)
Adds keybinds to the game
- A = Attack
- E = End Turn
- C = Chat
- H = Show/Hide tooltips
- M = Mute

This PR also makes chat take up the entire screen on mobile. It should
be more usable now.
C, H, M are shown on the buttons, but im not sure where to document A
and E.

Muting has been updated to be another check on top of the volume checks.
This makes the mute key easier to work with.