Change Strengths to readonly and reorganize Expansions (#760)
With these changes, the Strengths dict and the Expansions dict are both
read only. This is important so people don't accidentally think they are
supposed to modify these.
In addition, the expansions are now set up in a more intuitive way, and
there also shouldn't be typing issues with them anymore, like there were
in the past.