Add Viewer Component (#706)
This PR adds the viewer component.
The viewer component is a component that represents somebody who is
looking at the game. This can be a player for the game or a spectator.
Games will not need viewers to run which will make it possible to test
card interactions as unittests.