Add Types To Redux (#807)
This PR adds types to redux. I added types for the client local messages
and communication between the client and the server. I additionally
refactored the server to use selectors because that makes the control
flow easier to follow.


This pull request is NOT intended to be a final version of all of this.
This part of the code will need to be reorganized and refactored. The
goal of this PR is to add types to make that doable.