Fix Evil X Disabling Secondary Attacks At Incorrect Times (#860)
Evil X would disable his own secondary attack when it could not disable
an opponents attack. He also did this when he was not active. This
behavior does not match with other cards in the game so it now does not
remove the option for the attack and only removes the ability.