Implement Component Tetsts, Fix Battle Log, Fix Matchmaking Saga (#938)
I apologize for making such a large PR but I kinda got into it while
working on this. This PR fixes visual issues with the chat log, fixes a
bug where messages would be sent twice for players and spectators that
join before the game starts, and adds tests.

## IMPORTANT
`debug-config.json` is now `debug-config.js`. This is due to
playwright's component testing tools not liking json imports.