Fix minor issues with tag system (#962)
I'll be honest, this is not my best code. But HTML behavior is extremely
strange sometimes, and at least this definitely works.

- Deck List updates properly when duplicating, importing, or deleting
decks
- Tag filter works properly when loading into the screen
- Tag color doesn't change back to white after creating a tag. This was
intended, but I'm not really sure why I decided to do this. Anyway, the
old way I implemented it was wrong anyway so I fixed it.
- The deck screen ALWAYS scrolls to the correct position, so the deck
you selected is always visible. New feature, but this fixes some issues
too.