[Dead Reckoning 2] Split game away from concept of viewers (#1156)
The idea of this PR is to get the things for "running the game" outside
of the game object and into something else. Removing the things
necessary for player connections will make chat easier to manage and
will allow the game to be run on the client for dead reckoning in the
future.
I also removed viewers for the same reason. It also had the side effect
of making testing simpler too in my opinion.

**Bug Fixes**
Fixes a bug where chat would not be updated when a player reconnects to
a game.

**Future Work**
A future plan is to make `GameController` into an ABC. This would allow
both the server and the client to implement what they need to get the
game running on their end and we can share the actual implementation.