Active decks rework (#1132)
Instead of the active deck of a player being the full deck information,
the active deck is now only saved/selected as the deck code. This means
that it can be selected instantly and not when the deck info is
retrieved from the database. When a player queues into matchmaking, the
full deck info is retrieved from the database. This ensures players are
never matched into games with the wrong deck.

**Bug Fixes**
- Fixes the deck you queue into a game with sometimes being outdated if
you edited the deck on another device.
- Fixes the wrong deck being used when players with bad ping quickly
queue into a game after changing their deck.