Implement Basic Fuzz Testing (#1189)
This is a draft for fuzzy testing games. The idea is to play two AIs
against each other in a bunch of game and verify that the game never
gets into a valid state. For example, there should NEVER be a pick
request with no options.
The second goal is to verify that asserts never trigger. Asserts are
great, but only when you have them along with fuzzing.

This should be a huge help to the game's stability once it is
implemented.

**Bug Fixes**
Looting now asks for you to pick a card before attacking. This is to fix
looting creating a pick request with no choices after the user knocks
out the opponent's active hermit.