Increase fuzz-test amount to 500 (#1202)
So far it has been successful, so I want to increase the amount. The
amount of time this should take still seems like it will be relatively
low so I don't feel like we are wasting water/energy for this. Kinda a
funny consideration but it is worth it when you are blasting these
things lol.