Rework session-saga.ts, and database functions (#1235)
This PR updates session saga to allow for players to be automatically
logged in.
Players are now added using the API, and sockets always have a player
UUID.
In addition, I'm reworking a lot of the database functions to be neater
and work better with the new way players are added to the database.
Also adds icons to starter decks.
Font end changes to the login screen have not been implemented in this PR