Optimize Slot Empty Query (#1339)
This PR optimizes the `slot.empty` query to run in `O(1)` time. Now that
we are getting more players, I am finally looking back to optimizing
this system at the cost of complexity.
Note that now we can't link multiple cards to one slot, but we didn't do
that a single time since the system was made so I seriously doubt this
is important.