Low Hanging Fruit Game Performance Optimizations (#1344)
This PR reduces renrenders of the game board. It seemed to make modals
open a little faster for me but i could be imagining that. It will
definitely reduce power consumption though.

## Further Work
Improve the way status effects are read to reduce rerenders of the game
board when a status effect is changed.