Tag fix (#1342)
requires #1341 because i couldn't work on this without that being done 
Actually was not a bug here, the system was just unintuitive. I clear
the color and name to make it more clear that "add" will create a tag
with a different key when you click it, even if the name and color are
the same.

---------

Co-authored-by: Lunarmagpie <Bambolambo0@gmail.com>